fix sidebar
