+++ apps/theme/components/OverviewComponents/OverviewComponents.tsx
@@ -4,15 +4,13 @@ import {
   Heading,
   Link,
   Paragraph,
-  Search,
-  Select,
-  Table,
   Tag,
   Textfield,
 } from '@digdir/designsystemet-react';
 import { forwardRef } from 'react';
 import classes from './OverviewComponents.module.css';
 import { SettingsCard } from './SettingsCard/SettingsCard';
+import { TableCard } from './TableCard/TableCard';
 
 const users = [
   {
@@ -29,27 +27,6 @@ const users = [
     name: 'Marcus Viken',
     role: 'Backend',
     avatar: 'img/avatars/male3.png',
-  },
-];
-
-const tableData = [
-  {
-    name: 'Ola Normann',
-    avatar: 'img/avatars/male2.png',
-    email: 'ola@example.no',
-    phone: '12345678',
-  },
-  {
-    name: 'Kari Slotsveen',
-    avatar: 'img/avatars/female2.png',
-    email: 'kari@example.no',
-    phone: '12345678',
-  },
-  {
-    name: 'Marcus Viken',
-    avatar: 'img/avatars/male3.png',
-    email: 'marcus@example.no',
-    phone: '12345678',
   },
 ];
 
@@ -76,63 +53,7 @@ export const OverviewComponents = forwardRef<HTMLDivElement>(
               flexGrow: 1,
             }}
           >
-            <div style={{ display: 'flex', alignItems: 'center', gap: '12px' }}>
-              <Select data-size='sm' aria-label='Velg handling'>
-                <Select.Option value='blank'>Velg handling</Select.Option>
-                <Select.Option value='everest'>Dupliser</Select.Option>
-                <Select.Option value='aconcagua'>Slett</Select.Option>
-                <Select.Option value='denali'>Oppdater</Select.Option>
-              </Select>
-              <Button className={classes.tableBtn} data-size='sm'>
-                Utfør
-              </Button>
-            </div>
-            <Search className={classes.tableSearch} data-size='sm'>
-              <Search.Input
-                aria-label='Søk'
-                placeholder='Søk etter brukere her...'
-              />
-              <Search.Clear />
-            </Search>
-            <Table data-size='sm' className={classes.table}>
-              <Table.Head>
-                <Table.Row>
-                  <Table.HeaderCell onClick={function Ya() {}} sort='none'>
-                    Navn
-                  </Table.HeaderCell>
-                  <Table.HeaderCell>Epost</Table.HeaderCell>
-                  <Table.HeaderCell onClick={function Ya() {}} sort='none'>
-                    Telefon
-                  </Table.HeaderCell>
-                </Table.Row>
-              </Table.Head>
-              <Table.Body>
-                {tableData.map((user) => {
-                  return (
-                    <Table.Row key={user.name}>
-                      <Table.Cell
-                        style={{
-                          display: 'flex',
-                          alignItems: 'center',
-                          gap: '12px',
-                        }}
-                      >
-                        <Avatar aria-label={user.name} variant='square'>
-                          <img
-                            className={classes.tableImg}
-                            src={user.avatar}
-                            alt=''
-                          />
-                        </Avatar>
-                        {user.name}
-                      </Table.Cell>
-                      <Table.Cell>{user.email}</Table.Cell>
-                      <Table.Cell>{user.phone}</Table.Cell>
-                    </Table.Row>
-                  );
-                })}
-              </Table.Body>
-            </Table>
+            <TableCard />
           </div>
           <div className={classes.card} data-size='sm'>
             <SettingsCard />
@@ -154,7 +75,9 @@ export const OverviewComponents = forwardRef<HTMLDivElement>(
             </div>
           </div>
           <div className={classes.card} style={{ flexGrow: 1 }}>
-            <Heading data-size='2xs'>Folk du kanskje kjenner</Heading>
+            <Heading data-size='xs' level={3}>
+              Folk du kanskje kjenner
+            </Heading>
             <div className={classes.users}>
               {users.map((user) => {
                 return (
