+++ apps/theme/components/Sidebar/Sidebar.module.css
@@ -11,10 +11,16 @@
   flex-grow: 1;
   border-radius: 8px;
   margin-top: -170px;
-  min-height: calc(81vh);
-  max-height: calc(100vh - 40px);
+  height: calc(100vh - 200px);
   display: flex;
   flex-direction: column;
+}
+
+.sidebar > div:first-child {
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+  min-height: 0;
 }
 
 .showSidebar {
@@ -31,38 +37,6 @@
   display: flex;
   flex-direction: column;
   justify-content: space-between;
-  overflow: auto;
-  padding-bottom: 91px;
-}
-
-.sticky {
-  position: fixed;
-  margin-top: 0;
-}
-
-.title {
-  margin-bottom: 20px;
-}
-
-.group {
-  margin-bottom: 24px;
-  margin-top: -8px;
-}
-
-.group:last-child {
-  margin-bottom: 0;
-}
-
-.groupHeader {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 8px;
-  height: 42px;
-}
-
-.AddBtn {
-  margin-right: -8px;
 }
 
 .error {
@@ -88,13 +62,11 @@
 }
 
 .bottom {
-  position: absolute;
-  bottom: 0;
-  right: 0;
-  left: 0;
   border-radius: 0 0 8px 8px;
   padding: 24px;
   border-top: 1px solid var(--ds-color-neutral-border-subtle);
+  background-color: var(--ds-color-neutral-background-default);
+  flex-shrink: 0;
 }
 
 .appearance {
@@ -128,4 +100,15 @@
   .toggleOpen {
     right: 440px;
   }
+}
+
+.tabContent {
+  flex: 1;
+  min-height: 0;
+}
+
+.tabPanel {
+  height: 100%;
+  overflow-y: auto;
+  padding: 24px;
 }
